Clean up some command edges
Fix importing
