+++ cmd/configCmd.go
@@ -3,10 +3,9 @@ package cmd
 import (
 	"fmt"
 	"os"
-	"os/exec"
 
+	"github.com/Jefferson-Faseler/prion/internal/vimconfig"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
 var configCmd = &cobra.Command{
@@ -27,11 +26,11 @@ var configAddCmd = &cobra.Command{
 		configurationText := args[0]
 		fmt.Println("Adding " + configurationText + " to vimrc")
 
-		vimrc, err := os.OpenFile(vimrcPath(), os.O_APPEND|os.O_WRONLY, 0666)
+		vimrc, err := vimconfig.PrepareForWrite()
 		handleError(err)
 
 		defer vimrc.Close()
-		_, err = vimrc.WriteString(configurationText)
+		err = vimrc.Write(configurationText)
 		handleError(err)
 	},
 }
@@ -40,28 +39,11 @@ var configAddCmd = &cobra.Command{
 var configEditCmd = &cobra.Command{
 	Use:   "edit",
 	Short: "edit your vim configuration manually",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		editor := os.Getenv("EDITOR")
-		if editor == "" {
-			editor = "vim" // I mean, this makes sense, right?
-		}
-
-		fmt.Println("Opening editor for manual changes. Using editor " + editor)
-		fmt.Println("Close editor to continue")
-		executable, err := exec.LookPath(editor)
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("Opening editor for changes.")
+		err := vimconfig.EditVimRC()
 		handleError(err)
-
-		editorCmd := exec.Command(executable, vimrcPath())
-		editorCmd.Stdin = os.Stdin
-		editorCmd.Stdout = os.Stdout
-		editorCmd.Stderr = os.Stderr
-
-		return editorCmd.Run()
 	},
-}
-
-func vimrcPath() string {
-	return viper.GetString("VIMRC_PATH")
 }
 
 func init() {
